fix submodule clone url
The previous clone URL for scorep_plugin_cxx_wrapper required SSH with
standard configuration for github to clone. Subsequently, if
`ssh git@github.com` is not configured on the machine used for the
clone, the fetching of submodules fails.

This commit changes the clone URL to the public https:// version of the
repository.